rag/requirements.txt
chromadb==0.5.0
sentence-transformers==3.0.0
python-dotenv==1.0.1
pyyaml==6.0.1
numpy==1.26.4